+++ NEWS.md
@@ -2,7 +2,10 @@
 
 # development version
 
-- Add Kim 2023 camostat mesylate ([doi:10.3390/pharmaceutics15092357](https://doi.org/10.3390/pharmaceutics15092357)) -- healthy Korean adult men.
+- Add Leber 2023 tacrolimus, ciclosporin, mycophenolate mofetil, everolimus and
+  methylprednisolone during CytoSorb hemoperfusion
+  ([doi:10.3389/fmed.2023.1258661](https://doi.org/10.3389/fmed.2023.1258661)) --
+  healthy sheep on an extracorporeal circuit.
 
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.